Move cog to `__init__.py`
